Move `Read` traits to `store`
