+++ libs/@local/graph/graph/src/store/postgres/ontology/property_type.rs
@@ -20,7 +20,7 @@ use hash_graph_store::{
         GetPropertyTypesResponse, PropertyTypeQueryPath, PropertyTypeStore,
         UnarchivePropertyTypeParams, UpdatePropertyTypeEmbeddingParams, UpdatePropertyTypesParams,
     },
-    query::VersionedUrlSorting,
+    query::{QueryResult as _, Read as _, ReadPaginated, VersionedUrlSorting},
     subgraph::{
         Subgraph, SubgraphRecord as _,
         edges::{EdgeDirection, GraphResolveDepths, OntologyEdgeKind},
@@ -48,7 +48,6 @@ use type_system::{
 };
 
 use crate::store::{
-    crud::{QueryResult as _, Read as _, ReadPaginated},
     error::DeletionError,
     postgres::{
         AsClient, PostgresStore, TraversalContext,
